Setup the common and dashboard static directories
We want to make it difficult for users to access the dashboard.
The simplest way to do that is separate the common static files
from the sensitive files associated with the dashboard.

This is accomplished by setting the server up to serve all files
in the `frontend/common` directory relative to the `/` route and
files in the `frontend/dashboard` directory relative to a randomly
generated hexadecimal string.

The dashboard string is randomly generated on instantiation of the
server, so it will be different every time sRAT is run.