Fix multiple calls to res.send crash
The problem was that all of the nested callbacks in the `add_quiz`
function were calling `res.send` at some point. This caused an
error.

We fixed this by liberal use of Promises to ensure that errors are
getting passed up the call stack to the top level where we can
finally call `res.send` ONCE.